+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.121
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chainlink/token-allocation-adapter@1.14.8
+
 ## 3.0.120
 
 ### Patch Changes